Se realizó una mejora en el parametro dinero ingresado.

+++ src/main/kotlin/StateMachine.kt
@@ -34,7 +34,7 @@ object StateMachine {
         currentState.onEnter(this)
     }
 
-    fun solicitarPago(amount: Int) {
+    fun solicitarPago(amount: Double) {
         if (currentState is CoffeeMachineState.Idle) {
             if (amount >= 1) {
                 println("Pago recibido: $amount euro(s). Preparando café...")
